chore: split useQuotes
