+++ src/pages/quotes/QuoteDetailsVersions.tsx
@@ -8,7 +8,7 @@ import { useOrganizationInfos } from '~/hooks/useOrganizationInfos'
 
 import { getQuoteStatusMapping } from './common/getQuoteStatusMapping'
 import { getQuoteTypeTranslationKey } from './common/getQuoteTypetranslationKey'
-import { useQuotes } from './useQuotes'
+import { useQuotes } from './hooks/useQuotes'
 
 interface QuoteDetailsVersionsProps {
   quote: QuoteDetailItemFragment
