+++ src/pages/quotes/hooks/useQuotes.ts
@@ -1,11 +1,7 @@
 import { gql } from '@apollo/client'
 import { useEffect } from 'react'
 
-import {
-  GetQuotesQueryVariables,
-  useGetQuoteQuery,
-  useGetQuotesLazyQuery,
-} from '~/generated/graphql'
+import { GetQuotesQueryVariables, useGetQuotesLazyQuery } from '~/generated/graphql'
 
 gql`
   fragment QuoteListItem on Quote {
@@ -50,29 +46,6 @@ gql`
   }
 `
 
-gql`
-  fragment QuoteDetailItem on Quote {
-    id
-    number
-    status
-    version
-    orderType
-    currency
-    createdAt
-    customer {
-      id
-      name
-      externalId
-    }
-  }
-
-  query getQuote($id: ID!) {
-    quote(id: $id) {
-      ...QuoteDetailItem
-    }
-  }
-`
-
 export const useQuotes = (variables?: Omit<GetQuotesQueryVariables, 'limit' | 'page'>) => {
   const [getQuotes, { data, loading, error, fetchMore }] = useGetQuotesLazyQuery({
     variables: {
@@ -94,18 +67,5 @@ export const useQuotes = (variables?: Omit<GetQuotesQueryVariables, 'limit' | 'p
     loading,
     error,
     fetchMore,
-  }
-}
-
-export const useQuote = (id?: string) => {
-  const { data, loading, error } = useGetQuoteQuery({
-    variables: { id: id || '' },
-    skip: !id,
-  })
-
-  return {
-    quote: data?.quote,
-    loading,
-    error,
   }
 }
